TASK: Move NodeCreation stuff around
