+++ Classes/Domain/NodeCreation/NodeCreationCommands.php
@@ -1,6 +1,6 @@
 <?php
 declare(strict_types=1);
-namespace Neos\Neos\Ui\NodeCreationHandler;
+namespace Neos\Neos\Ui\Domain\NodeCreation;
 
 /*
  * This file is part of the Neos.Neos.Ui package.
